Replace arbitrary tuples by functions (naïve implementation).
This implementation has two problems:

1. Since type identifiers are associated to function application we
   loose polymorphism. This is particularly annoying since we need to
   instantiate e.g. int lists from string lists.

2. The function lifting combinators reverse the arguments of the
   function. 'a -> 'b -> 'c becomes 'b -> 'a -> 'c. It's not
   exceedingly annoying but a bit nonetheless.

We should be able to overcome 1. by pushing the type identifiers
to the constants and variables of the term. The annoying thing
is that we'll need to type our variables (i.e. add an argument
to `fresh`).

We still can't play far with the system because of the finite depth
first search.